Remove: Consolidated Spider
To initiate a scrape, you must return a “request object” back to scrapy
for it to do its work. This prevents the calling function to get data
back from a scrape at the point where it was called, which is necessary
for pages that require additional scraping from 1 or more sites.